Rearrange files, output cov
