+++ translator_template/tests/test_parse.py
@@ -1,4 +1,5 @@
 from io import StringIO
+import pytest
 
 from translator_template.parse import read
 
@@ -10,6 +11,7 @@ def test_read():
 
     assert [("PAGE 1", 2)] == read_from("2")
     assert [("PAGE 1", 2), ("PAGE 2", 3)] == read_from("23")
+    assert [("PAGE 1", 2), ("PAGE 2", 3)] == read_from("2 3")
     assert [("PAGE 1", 2), ("PAGES 2-3", 3)] == read_from("23-")
     assert [("PAGE 1", 56)] == read_from("*56*")
     assert [("PAGES 1-2", 56), ("PAGE 3", 4)] == read_from("*56*-4")
@@ -17,3 +19,8 @@ def test_read():
     assert [("PAGE 1", 1), ("PAGE 2", 23), ("PAGE 3", 4)] == read_from("1*23*4")
     assert [("PAGE 1", 1), ("PAGE 2", 23), ("PAGE 3", 45)] == read_from("1*23**45*")
     assert [("PAGE 1", 1), ("PAGES 2-3", 23), ("PAGE 4", 45)] == read_from("1*23*-*45*")
+
+
+def test_read_bad_input():
+    with pytest.raises(RuntimeError):
+        next(read(StringIO("X")))
